Add untested message box and tutorial system
Tutorial system based off of the idea for tutorials to be less annoying
and in-your-face. Basically we want to display a helpful message if and
only if the user is getting stuck and/or trying to do something that is
sub-optimal (deleting a node instead of deleting an edge).

In those cases, we want to display a message on screen of the
step-by-step instructions to do something.

Ref: #4