Implement #8
Kind-of fix dialog system. Since we can only have 1 dialog box at a
time, and since there is no way to queue up dialog boxes, we have to
asynchronously delay the calling of `install` on dialog boxes for
tutorials in the off chance that (like for the templating tutorial) we
have to trigger a tutorial directly after a dialog box.

Currently set to trigger 500ms after the function `trigger()` is called.
Some may call this a hotfix. I would call it just another day. Will
refactor if the current way of doing things no longer works.